requirements.txt
chromadb
requests
mysql-connector-python
fastapi
uvicorn[standard]
pydantic
ollama
flask
gunicorn
mcp[cli]
